yolo-executor/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

# Base ----------------------------------------
matplotlib==3.7.1
numpy==1.23.5
opencv-python==4.7.0.72
Pillow==9.5.0
PyYAML==6.0
requests==2.28.1
httpx==0.24.1
scipy==1.10.1
torch==1.10.2+cpu # 1.10.2+cpu if no GPU, 1.10.2+cu113 for CUDA 11.3
torchvision==0.11.3+cpu # 0.11.3+cpu if no GPU, 0.11.3+cu113 for CUDA 11.3
tqdm==4.65.0
protobuf==4.21.2

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Server --------------------------------------
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
